+++ src/components/containers/ContainerTask/index.tsx
@@ -13,7 +13,9 @@ import * as S from "./styles";
 interface IContainerTaskProps {
   setIsTaskOpen: React.Dispatch<React.SetStateAction<boolean>>;
   setIsDeleteTaskOpen: React.Dispatch<React.SetStateAction<boolean>>;
-  setCrudTasksOptions: React.Dispatch<React.SetStateAction<"addTask" | "editTask" | "deleteTask" | null>>;
+  setCrudTasksOptions: React.Dispatch<
+    React.SetStateAction<"addTask" | "editTask" | "deleteTask" | "duplicateTask" | null>
+  >;
 }
 
 export enum Iterator {
